add url to highest level node in the tree since it was missing
